requirements.txt
fastapi>=0.115
uvicorn>=0.30
pandas>=2.2
numpy>=1.26
sentence-transformers>=3.0
faiss-cpu>=1.8
python-dotenv>=1.0
requests>=2.32
pydantic>=2.7
scikit-learn>=1.5
python-multipart>=0.0.9
